% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_P.R
\docType{data}
\name{A_P}
\alias{A_P}
\title{Ocean pre-fishery abundance}
\format{
\subsection{\code{A_P}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/do-run-reconstruction
}
\usage{
A_P
}
\description{
Ocean pre-fishery abundance, which is the preterminal post-fishery
abundance plus the fish caught in preterminal non-net fisheries (e.g., troll).
}
\keyword{datasets}
