% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B-star.R
\docType{data}
\name{B_star}
\alias{B_star}
\title{Female brood removals by age}
\format{
\subsection{\code{B_star}}{

An array with two dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/process-B-star.R
}
\usage{
B_star
}
\description{
Brood removals in the Kitsumkalum River, females only, by age
}
\keyword{datasets}
