% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\docType{data}
\name{E}
\alias{E}
\title{Escapement by population and year}
\format{
\subsection{\code{E}}{

An array with two dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
}
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
E
}
\description{
Escapement of Chinook by population and year.
}
\keyword{datasets}
