% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_star.R
\docType{data}
\name{E_star}
\alias{E_star}
\title{Escapement by population, year, and age}
\format{
\subsection{\code{E_star}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
E_star
}
\description{
Escapement of Chinook by population, year, and age.
}
\keyword{datasets}
