% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA-data-processed.R
\docType{data}
\name{ERA_data_processed}
\alias{ERA_data_processed}
\title{Summary table of variables derived from Chinook Technical Committee Exploitation Rate Analysis data}
\format{
\subsection{\code{ERA_data_processed}}{

List with two elements:

Element one: data frame with columns:
\describe{
\item{y}{Year}
\item{a}{Age}
\item{H_star}{Hatchery origin spawners}
\item{tau_dot_M}{Terminal marine exploitation rate}
\item{phi_dot_M}{Preterminal net harvest rate of mature fish}
\item{r}{Maturation rate}
\item{phi_dot_E}{Preterminal exploitation rate (non-net fisheries)}
\item{Q}{Adult equivalency rate}
}

Element two: same data as element one but long format.
}
}
\source{
data-raw/kitsumkalum/process-ERA-data.R
}
\usage{
ERA_data_processed
}
\description{
Data frames (one wide format, one long format) of all the variables derived from CTC ERA output data.
}
\keyword{datasets}
