% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G.R
\docType{data}
\name{G}
\alias{G}
\title{Gill-net catch data}
\format{
\subsection{\code{G}}{

An array with two dimensions:
\describe{
\item{w}{Week}
\item{y}{Year}
}
}
}
\source{
data-raw/process-G.R
}
\usage{
G
}
\description{
Gill-net test fishery data by week and year. Includes adults (age 4+)
}
\keyword{datasets}
