% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.R
\docType{data}
\name{H}
\alias{H}
\title{Number of hatchery origin Kitsumkalum Chinook in escapement, by return year}
\format{
\subsection{\code{H}}{

Numeric vector, with names of return year.
}
}
\source{
data-raw/kitsumkalum/process-ERA-data.R
}
\usage{
H
}
\description{
From the Chinook CTC ERA model outputs, number of fish with CWTs in escapement,
expanded by mark rate.
}
\keyword{datasets}
