% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H-star.R
\docType{data}
\name{H_star}
\alias{H_star}
\title{Hatchery origin Chinook in escapement, Kitsumkalum River, by return year and age}
\format{
\subsection{\code{H_star}}{

An array with two dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/kitsumkalum/process-ERA-data.R
}
\usage{
H_star
}
\description{
From the Chinook CTC ERA model outputs, number of fish with CWTs in escapement,
expanded by mark rate.
}
\keyword{datasets}
