% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K-star.R
\docType{data}
\name{K_star}
\alias{K_star}
\title{Kitsumkalum escapement data by year and age}
\format{
\subsection{\code{K_star}}{

An array with two dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/kitsumkalum/process-K-star.R, skrunchy::get_K_star(), data-raw/kitsumkalum/kitsumkalum-escapement-by-age-1984-2020.xlsx
}
\usage{
K_star
}
\description{
Kitsumkalum escapement data by year and age
}
\keyword{datasets}
