% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitsumkalum.R
\docType{data}
\name{kitsumkalum}
\alias{kitsumkalum}
\title{Kitsumkalum escapement data by year}
\format{
\subsection{\code{kitsumkalum}}{

A dataframe with three variables:
\describe{
\item{year}{Year}
\item{kitsumkalum_escapement}{Escapement of Kitsumkalum Chinook ages 4-7, no jacks, estimated using mark-recapture and POPAN.}
\item{sd}{Standard Deviation of kitsumkalum_escapement.}
}
}
}
\source{
data-raw/process-kitsumkalum.R, data-raw/kitsumkalum-escapement.csv
}
\usage{
kitsumkalum
}
\description{
Kitsumkalum escapement data by year
}
\keyword{datasets}
