% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatureRun.R
\docType{data}
\name{MatureRun}
\alias{MatureRun}
\title{Mature Run}
\format{
\subsection{\code{MatureRun}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
MatureRun
}
\description{
Mature run by population, year, and age
}
\keyword{datasets}
