% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N.R
\docType{data}
\name{N}
\alias{N}
\title{Total run}
\format{
\subsection{\code{N}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
N
}
\description{
Sum of escapement, brood removals, terminal mortality, and preterminal mortality (including incidental mortality).
}
\keyword{datasets}
