% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_age_kitsumkalum.R
\docType{data}
\name{n_age_kitsumkalum}
\alias{n_age_kitsumkalum}
\title{Number of age observations for Kitsumkalum spawners}
\format{
\subsection{\code{n_age_kitsumkalum}}{

Array with 3 dimensions:
\describe{
\item{s}{Sex}
\item{y}{Return year}
\item{a}{Total age (Gilbert-Rich)}
}
}
}
\source{
data-raw/process-kitsumkalum-age-observations-omega-K.R
}
\usage{
n_age_kitsumkalum
}
\description{
Number of Kitsumkalum spawners by age, year, and sex. From scales collected on the
spawning grounds during tagging and dead pitch.
}
\keyword{datasets}
