% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-age-kitsumkalum.R, R/n-age-observations.R
\docType{data}
\name{n_age_observations}
\alias{n_age_observations}
\title{Number of age observations by sex, year, and age for Kitsumkalum.}
\format{
\subsection{\code{n_kitsumkalum}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}

\subsection{\code{n_age_observations}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/kitsumkalum/process-kitsumkalum-age-observations-omega-K.R

data-raw/process-omega-and-biodata.R
}
\usage{
n_age_observations

n_age_observations
}
\description{
Used to get omega_K, age proportions for Kitsumkalum River. Which is used to get K_star, age-specific Kitsumkalum escapement.

Array of number of age observations at Tyee Test Fishery.
}
\keyword{datasets}
