% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\docType{data}
\name{omega}
\alias{omega}
\title{Age proportion data by population, year, and age}
\format{
\subsection{\code{omega}}{

List with two objects

\describe{
\item{omega}{Array}
\item{df}{Data frame}
}
}
}
\source{
data-raw/process-omega-and-biodata.R
}
\usage{
omega
}
\description{
Age proportion by population, year, and age, for ages 4-7. From Tyee Test Fishery.
}
\keyword{datasets}
