% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_J.R
\docType{data}
\name{omega_J}
\alias{omega_J}
\title{Age proportion data by population, year, and age, including jacks}
\format{
\subsection{\code{omega_J}}{

List with two objects

\describe{
\item{omega}{Array}
\item{df}{Data frame}
}
}
}
\source{
data-raw/process-omega-and-biodata.R
}
\usage{
omega_J
}
\description{
Age proportion by population, year, and age, for ages 3-7. From Tyee Test Fishery.
}
\keyword{datasets}
