% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_K.R
\docType{data}
\name{omega_K}
\alias{omega_K}
\title{Age proportions of Kitsumkalum spawners by sex}
\format{
\subsection{\code{omega_K}}{

Array with three dimensions:
\describe{
\item{s}{Sex}
\item{y}{Return year}
\item{a}{Age}
}
}
}
\source{
data-raw/process-kitsumkalum-age-observations-omega-K.R
}
\usage{
omega_K
}
\description{
Age proportions of Kitsumkalum spawners (collected on spawning grounds) by year, sex, and age. From scales collected during tagging and deadpitch.
}
\keyword{datasets}
