% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P.R
\docType{data}
\name{P}
\alias{P}
\title{Genetic proportion data}
\format{
\subsection{\code{P}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{w}{Week}
\item{y}{Year}
}
}
}
\source{
data-raw/process-P.R
}
\usage{
P
}
\description{
Genetic proportion data
}
\keyword{datasets}
