% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_proportion_wild.R
\docType{data}
\name{p_proportion_wild}
\alias{p_proportion_wild}
\title{Proportion wild spawners}
\format{
\subsection{\code{p_proportion_wild}}{

List with the data in two formats:
\describe{
\item{p}{Array}
\item{df}{Data frame}
}
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
p_proportion_wild
}
\description{
Proportion of spawners that are wild origin
}
\keyword{datasets}
