% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_tilde.R
\docType{data}
\name{P_tilde}
\alias{P_tilde}
\title{Annual genetic proportions at Tyee}
\format{
\subsection{\code{P_tilde}}{

List of 3 elements. First is an array of pooled genetic proportions with 2 dimensions: i (population) and y (year). Second is an array of SD of pooled genetic proportions with 2 dimensions: i (population) and y (year). Third is a data frame with P_tilde, sigma_P_tilde, and year merged for plotting and reporting.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
P_tilde
}
\description{
Genetic proportions at Tyee after expanding weekly proportions by catch
}
\keyword{datasets}
