% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_dot_E.R
\docType{data}
\name{phi_dot_E}
\alias{phi_dot_E}
\title{Preterminal total mortality exploitation rate for non-net fisheries}
\format{
\subsection{\code{phi_dot_E}}{

Array with two dimensions: y (year) and a (age).
}
}
\source{
data-raw/kitsumkalum/process-ERA-data.R
}
\usage{
phi_dot_E
}
\description{
Exploitation rate for preterminal non-net fisheries (e.g., troll). From the CTC Chinook model output for Kitsumkalum CWT releases (KLM). Equal to the number of fish caught plus incidental mortalities for non-net fisheries (e.g., troll), in areas seaward of the terminal area, divided by total stock for the cohort, in nominal fish (not adjusted for AEQs).
}
\keyword{datasets}
