% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_N.R
\docType{data}
\name{phi_N}
\alias{phi_N}
\title{Preterminal fishing mortality in nominal fish}
\format{
\subsection{\code{phi_N}}{

List with two elements. First element: numeric, array of preterminal fishing mortality in nominal fish, with three dimensions: population (i), year (y), and age (a). Second element: numeric, dataframe version of first element.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
phi_N
}
\description{
The number of fish caught in pre-terminal non-net fisheries (e.g., troll).
}
\keyword{datasets}
