% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_Q.R
\docType{data}
\name{phi_Q}
\alias{phi_Q}
\title{Preterminal fishing mortality in adult equivalents}
\format{
\subsection{\code{phi_Q}}{

List with two elements. First element: numeric, array of preterminal fishing mortality in adult equivalents, with three dimensions: population (i), year (y), and age (a). Second element: dataframe version of first element, for plotting and tables.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
phi_Q
}
\description{
Preteriminal fishing mortality in adult equivalents (AEQ).
}
\keyword{datasets}
