% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.R
\docType{data}
\name{Q}
\alias{Q}
\title{Adult Equivalency (AEQ) rates}
\format{
\subsection{\code{Q}}{

array of adult equivalency (AEQ) rates with two dimensions: year (y) and age (a). Output from CTC Chinook model based on Kitsumkalum CWT releases (KLM).
}
}
\source{
Source code to create object
}
\usage{
Q
}
\description{
Adult Equivalency (AEQ) rates by year and age.
}
\keyword{datasets}
