% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\docType{data}
\name{R}
\alias{R}
\title{Recruits}
\format{
\subsection{\code{R}}{

List with three elements. First element: numeric, array of recruits by  brood year with three dimensions: population (i), brood year (b).
Second element: data frame version of first element. Third element: data frame of recruits by age, population, brood year and return year.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
R
}
\description{
Recruits by brood year
}
\keyword{datasets}
