% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_maturation_rate.R
\docType{data}
\name{r_maturation_rate}
\alias{r_maturation_rate}
\title{Maturation rate}
\format{
\subsection{\code{r_maturation_rate}}{

Array of maturation rate by year and age. CTC Chinook model output for Kitsumkalum (KLM).
}
}
\source{
data-raw/kitsumkalum/process-ERA-data.R
}
\usage{
r_maturation_rate
}
\description{
Maturation rate by year and age.
}
\keyword{datasets}
