% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-reconstruction-table.R
\docType{data}
\name{run_reconstruction_table}
\alias{run_reconstruction_table}
\title{Big table with run reconstruction variables by population, return year, and age}
\format{
\subsection{\code{run_reconstruction_table}}{

Data frame with the following columns:
\describe{
\item{i}{Population}
\item{y}{Return year}
\item{a}{Age}
\item{...}{More variables}
}
}
}
\source{
Rmd/run-reconstruction-1984-2024
}
\usage{
run_reconstruction_table
}
\description{
Contains variables used in the run reconstruction.
}
\keyword{datasets}
