% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_reconstruction_table_summary.R
\docType{data}
\name{run_reconstruction_table_summary}
\alias{run_reconstruction_table_summary}
\title{Run reconstruction data summarized by return year}
\format{
\subsection{\code{run_reconstruction_table_summary}}{

Data frame
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
run_reconstruction_table_summary
}
\description{
Full run reconstruction data, summed across ages.
}
\keyword{datasets}
