% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_star.R
\docType{data}
\name{S_star}
\alias{S_star}
\title{Spawners by age}
\format{
\subsection{\code{S_star}}{

List with two elements. First element: numeric, array of spawner values with three dimensions: population (i), year (y), and age (a). Second element: data frame version of first element in long format, for plotting and tables.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
S_star
}
\description{
Total spawners
}
\keyword{datasets}
