% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma-P.R
\docType{data}
\name{sigma_P}
\alias{sigma_P}
\title{Standard error for genetic proportion data}
\format{
\subsection{\code{sigma_P}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{w}{Week}
\item{y}{Year}
}
}
}
\source{
data-raw/process-P.R
}
\usage{
sigma_P
}
\description{
Standard error of genetic proportion data by population, week, and year.
}
\keyword{datasets}
