% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_L.R
\docType{data}
\name{tau_L}
\alias{tau_L}
\title{Terminal mortality by age, lower Skeena}
\format{
\subsection{\code{tau_L}}{

A list with two objects. First object: numeric, array of total terminal mortalities in the lower Skeena with three dimensions: population (i), year (y), and age (a). Second object: data frame with same data in long format (columns for population, year, and age) for plotting.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
tau_L
}
\description{
Terminal freshwater mortality by age, lower Skeena (below Terrace)
}
\keyword{datasets}
