% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tau_L_total.R
\docType{data}
\name{Tau_L_total}
\alias{Tau_L_total}
\title{Terminal mortality, lower Skeena}
\format{
\subsection{\code{Tau_L_total}}{

A numeric vector of total terminal mortalities including catch and incidental mortalities, for the lower Skeena (downstream of Terrace), by year.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
Tau_L_total
}
\description{
Terminal mortality, lower Skeena (downstream of Terrace)
}
\keyword{datasets}
