% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_U.R
\docType{data}
\name{tau_U}
\alias{tau_U}
\title{Terminal mortality by age, upper Skeena}
\format{
\subsection{\code{tau_U}}{

A list with two objects. First object: numeric, array of total terminal mortalities in the upper Skeena with three dimensions: population (i), year (y), and age (a). Second object: data frame with same data in long format (columns for population, year, and age) for plotting.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
tau_U
}
\description{
Terminal mortality by age upstream of Terrace
}
\keyword{datasets}
