% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tau_U_total.R
\docType{data}
\name{Tau_U_total}
\alias{Tau_U_total}
\title{Terminal mortality, upper Skeena}
\format{
\subsection{\code{Tau_U_total}}{

A numeric vector of total terminal mortalities including catch and incidental mortalities, for the upper Skeena (upstream of Terrace), by year.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
Tau_U_total
}
\description{
Terminal mortality, upper Skeena (upstream of Terrace)
}
\keyword{datasets}
