% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_W.R
\docType{data}
\name{tau_W}
\alias{tau_W}
\title{Wild terminal mortalities by age}
\format{
\subsection{\code{tau_W}}{

List with two elements. First element: numeric, array of wild terminal mortalities with three dimensions: population (i), year (y), and age (a). Second element: data frame version of first element, for plotting and tables.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
tau_W
}
\description{
Wild terminal mortalities by age
}
\keyword{datasets}
