% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-name-key.R
\docType{data}
\name{variable_name_key}
\alias{variable_name_key}
\title{Table with key to variable names, skrunchy version and with more details.}
\format{
\subsection{\code{variable_name_key}}{

Data frame with the following columns:
\describe{
\item{skrunchy}{Variable names in skrunchy}
\item{detail}{More descriptive variable name}
\item{skrunchy_with_detail}{ First two columns combined together}
}
}
}
\source{
data-raw/variable-name-key.csv, data-raw/misc.R
}
\usage{
variable_name_key
}
\description{
Columns with skrunchy variable names and more descriptive names. Used to produce data outputs.
}
\keyword{datasets}
