% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_star.R
\docType{data}
\name{W_star}
\alias{W_star}
\title{Wild spawners by age}
\format{
\subsection{\code{W_star}}{

First element: numeric, array of wild spawner values with three dimensions: population (i), year (y), and age (a). Second element: data frame version of first element in long format, for plotting and tables.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
W_star
}
\description{
Wild (natural origin) spawners by age.
}
\keyword{datasets}
