% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.R
\docType{data}
\name{X}
\alias{X}
\title{Return to Terrace}
\format{
\subsection{\code{X}}{

A list with three elements. First element: Numeric, X which is an array of returns to Terrace with two dimensions: population (i) and year (y). Second element: numeric, sigma_X which is an array of SE of returns to Terrace of Chinook, with two dimensions: population (i) and year (y). Third element: dataframe with X, sigma_X, and year merged for plotting and reporting.
}
}
\source{
data-raw/do-run-reconstruction.R
}
\usage{
X
}
\description{
Returns to Terrace of adult Chinook salmon (X) for population i and year y. Uses pooled genetic proportions and the number of Chinook returns to the Kitsumkalum River.
}
\keyword{datasets}
