% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-age-observations.R
\docType{data}
\name{n_age_observations}
\alias{n_age_observations}
\title{Number of age observations by year, age, and CU}
\format{
\subsection{\code{n_age_observations}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/process-omega-and-biodata.R
}
\usage{
n_age_observations
}
\description{
Array of number of age observations at Tyee Test Fishery.
}
\keyword{datasets}
