% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B-star.R, R/P.R
\docType{data}
\name{P}
\alias{P}
\title{Female brood removals by age}
\format{
\subsection{\code{B_star}}{

An array with two dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}

\subsection{\code{P}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{w}{Week}
\item{y}{Year}
}
}
}
\source{
data-raw/process-B-star.R

data-raw/process-P.R
}
\usage{
P

P
}
\description{
Brood removals in the Kitsumkalum River, females only, by age

Genetic proportion data
}
\keyword{datasets}
